flask_cap/templates/requirements.txt
flask
flask_script
flask_restful
flask_sqlalchemy
click
gevent
gunicorn
oss2
pymysql
xlwt
pycrypto
rsa